ACC-280: update to ceiling the amount

+++ src/Classes/BasePayrollLiability.php
@@ -17,7 +17,7 @@ abstract class BasePayrollLiability
         $this->company_payroll = $company_payroll;
     }
 
-    abstract public function compute(Collection $tax_areas): float;
+    abstract public function compute(Collection $tax_areas): int;
 
-    abstract public function getWages(Collection $tax_areas): float;
+    abstract public function getWages(Collection $tax_areas): int;
 }
